[ENH] added save best model to inceptionTime classifier

+++ aeon/classification/deep_learning/inception_time.py
@@ -97,6 +97,24 @@ class InceptionTimeClassifier(BaseClassifier):
             list of tf.keras.callbacks.Callback objects.
         file_path           : str, default = "./"
             file_path when saving model_Checkpoint callback
+        save_best_model     : bool, default = False
+            Whether or not to save the best model, if the
+            modelcheckpoint callback is used by default,
+            this condition, if True, will prevent the
+            automatic deletion of the best saved model from
+            file and the user can choose the file name
+        save_last_model     : bool, default = False
+            Whether or not to save the last model, last
+            epoch trained, using the base class method
+            save_last_model_to_file
+        best_file_name      : str, default = "best_model"
+            The name of the file of the best model, if
+            save_best_model is set to False, this parameter
+            is discarded
+        last_file_name      : str, default = "last_model"
+            The name of the file of the last model, if
+            save_last_model is set to False, this parameter
+            is discarded
         random_state        : int, default = 0
             seed to any needed random actions.
         verbose             : boolean, default = False
@@ -150,6 +168,10 @@ class InceptionTimeClassifier(BaseClassifier):
         depth=6,
         use_custom_filters=True,
         file_path="./",
+        save_last_model=False,
+        save_best_model=False,
+        best_file_name="best_model",
+        last_file_name="last_model",
         batch_size=64,
         use_mini_batch_size=False,
         n_epochs=1500,
@@ -183,6 +205,11 @@ class InceptionTimeClassifier(BaseClassifier):
 
         self.file_path = file_path
 
+        self.save_last_model = save_last_model
+        self.save_best_model = save_best_model
+        self.best_file_name = best_file_name
+        self.last_file_name = last_file_name
+
         self.callbacks = callbacks
         self.random_state = random_state
         self.verbose = verbose
@@ -213,7 +240,7 @@ class InceptionTimeClassifier(BaseClassifier):
         self.classifers_ = []
         rng = check_random_state(self.random_state)
 
-        for _ in range(0, self.n_classifiers):
+        for n in range(0, self.n_classifiers):
             cls = IndividualInceptionClassifier(
                 nb_filters=self.nb_filters,
                 nb_conv_per_layer=self.nb_conv_per_layer,
@@ -230,6 +257,10 @@ class InceptionTimeClassifier(BaseClassifier):
                 depth=self.depth,
                 use_custom_filters=self.use_custom_filters,
                 file_path=self.file_path,
+                save_best_model=self.save_best_model,
+                save_last_model=self.save_last_model,
+                best_file_name=self.best_file_name + str(n),
+                last_file_name=self.last_file_name + str(n),
                 batch_size=self.batch_size,
                 use_mini_batch_size=self.use_mini_batch_size,
                 n_epochs=self.n_epochs,
@@ -375,7 +406,6 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
         use_custom_filters  : bool, default = True,
             condition on wether or not to use custom filters
             in the first inception module
-
         batch_size          : int, default = 64
             the number of samples per gradient update.
         use_mini_batch_size : bool, default = False
@@ -387,6 +417,24 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
             list of tf.keras.callbacks.Callback objects.
         file_path           : str, default = "./"
             file_path when saving model_Checkpoint callback
+        save_best_model     : bool, default = False
+            Whether or not to save the best model, if the
+            modelcheckpoint callback is used by default,
+            this condition, if True, will prevent the
+            automatic deletion of the best saved model from
+            file and the user can choose the file name
+        save_last_model     : bool, default = False
+            Whether or not to save the last model, last
+            epoch trained, using the base class method
+            save_last_model_to_file
+        best_file_name      : str, default = "best_model"
+            The name of the file of the best model, if
+            save_best_model is set to False, this parameter
+            is discarded
+        last_file_name      : str, default = "last_model"
+            The name of the file of the last model, if
+            save_last_model is set to False, this parameter
+            is discarded
         random_state        : int, default = 0
             seed to any needed random actions.
         verbose             : boolean, default = False
@@ -429,6 +477,10 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
         depth=6,
         use_custom_filters=True,
         file_path="./",
+        save_best_model=False,
+        save_last_model=False,
+        best_file_name="best_model",
+        last_file_name="last_model",
         batch_size=64,
         use_mini_batch_size=False,
         n_epochs=1500,
@@ -440,7 +492,9 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
         optimizer=None,
     ):
         _check_dl_dependencies(severity="error")
-        super(IndividualInceptionClassifier, self).__init__()
+        super(IndividualInceptionClassifier, self).__init__(
+            save_last_model=save_last_model, last_file_name=last_file_name
+        )
         # predefined
         self.nb_filters = nb_filters
         self.nb_conv_per_layer = nb_conv_per_layer
@@ -461,6 +515,11 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
         self.use_custom_filters = use_custom_filters
 
         self.file_path = file_path
+
+        self.save_best_model = save_best_model
+        self.save_last_model = save_last_model
+        self.best_file_name = best_file_name
+        self.last_file_name = last_file_name
 
         self.callbacks = callbacks
         self.random_state = random_state
@@ -486,7 +545,6 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
             bottleneck_size=self.bottleneck_size,
             depth=self.depth,
             use_custom_filters=self.use_custom_filters,
-            random_state=self.random_state,
         )
 
     def build_model(self, input_shape, n_classes, **kwargs):
@@ -569,7 +627,9 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
         if self.verbose:
             self.training_model_.summary()
 
-        self.file_name_ = str(time.time_ns())
+        self.file_name_ = (
+            self.best_file_name if self.save_best_model else str(time.time_ns())
+        )
 
         self.callbacks_ = (
             [
@@ -599,9 +659,13 @@ class IndividualInceptionClassifier(BaseDeepClassifier):
             self.model_ = tf.keras.models.load_model(
                 self.file_path + self.file_name_ + ".hdf5", compile=False
             )
-            os.remove(self.file_path + self.file_name_ + ".hdf5")
+            if not self.save_best_model:
+                os.remove(self.file_path + self.file_name_ + ".hdf5")
         except FileNotFoundError:
             self.model_ = deepcopy(self.training_model_)
+
+        if self.save_last_model:
+            self.save_last_model_to_file(file_path=self.file_path)
 
         return self
